split DefaultGameInstance into BaseGameInstance
