+++ src/main/java/work/lclpnet/ap2/game/cozy_campfire/CozyCampfireInstance.java
@@ -1,0 +1,21 @@
+package work.lclpnet.ap2.game.cozy_campfire;
+
+import work.lclpnet.ap2.api.game.MiniGameHandle;
+import work.lclpnet.ap2.impl.game.DefaultTeamGameInstance;
+
+public class CozyCampfireInstance extends DefaultTeamGameInstance {
+
+    public CozyCampfireInstance(MiniGameHandle gameHandle) {
+        super(gameHandle);
+    }
+
+    @Override
+    protected void prepare() {
+
+    }
+
+    @Override
+    protected void ready() {
+
+    }
+}
